% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{check_VC}
\alias{check_VC}
\title{Densitate de probabilitate pentru un obiect de tip v.a.continua}
\usage{
e_densitate(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{f}{o repartitie de v.a. continua}
}
\value{
Bool
}
\description{
Pt un obiect  se determina daca acesta este densitate de probabilitate
}
