% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{coef}
\alias{coef}
\title{Coefincientul de corelatie pentru 2 v.a. continue}
\usage{
coef(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)}, a, b, c, d)
}
\arguments{
\item{f, a, b, c, d}{: f-o repartitie de v.a. continua, \link{a,b}, \link{c,d} - intevalele de definitie ale celor 2 v.a continue}
}
\value{
number - coeficientul de corelatie, daca exista
}
\description{
Realizeaza denistatea comuna a 2 v.a. continue, apoi calculeaza Coefincientul de corelatie pe baza ei
}
