% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{constanta,VC-method}
\alias{constanta,VC-method}
\title{Constanta pentru obiect de v.a. continua}
\usage{
constanta(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\description{
Wrapper pt functia contanta a unui obiect de v.a. continua
Pt o functie f, intodusa de utlizator se determina o constanta de normalizare k
}
