% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{constanta}
\alias{constanta}
\title{Constanta de normalizare k}
\usage{
constanta(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{f}{o repartitie de v.a. continua}
}
\value{
number - constanta de normalizare daca exista
}
\description{
Pt o functie f, intodusa de utlizator se determina o constanta de normalizare k
}
