% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{d_exp}
\alias{d_exp}
\title{Densitate de probabilitate(Rep. Exponentiala)}
\usage{
d_exp(x, lambda)
}
\arguments{
\item{x, lamda}{: x - number si lambda}
}
\value{
number -  Densitate de probabilitate in Rep. Exponentiala
}
\description{
Calculeaza densitatea de probabilitate in x a unei v.a. continue in repartitia Exponentiala - functie folosita in repartizarea grafica
}
