% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{d_norm}
\alias{d_norm}
\title{Densitate de probabilitate(Rep.Normala)}
\usage{
d_norm(x, miu, sigma)
}
\arguments{
\item{x, miu, sigma_p}{: x - number si miu, sigma}
}
\value{
number -  Densitate de probabilitate in Rep. Normala
}
\description{
Calculeaza densitatea de probabilitate in x a unei v.a. continue in repartitia Normala - functie folosita in repartizarea grafica
}
