% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dispersie}
\alias{dispersie}
\title{Dispersia unei v.a continue X}
\usage{
dispersie(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{f:}{f-o repartitie de v.a. continua}
}
\value{
number - dispersia
}
\description{
Calculeaza dispersia unei v.a continue X
}
