% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dispersie2}
\alias{dispersie2}
\title{Dispersia unei v.a continue X trecuta print-o functie G}
\usage{
dispersie2(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)}, g(x){return (x+2)})
}
\arguments{
\item{f, g}{: f-o repartitie de v.a. continua, g functie continua}
}
\value{
number - dispersia
}
\description{
Calculeaza dispersia unei v.a continue X trecuta print-o functie G continua
}
