% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dpareto}
\alias{dpareto}
\title{Densitate de probabilitate(Rep.Pareto)}
\usage{
dpareto(x,m,alpha)
}
\arguments{
\item{x, m, alpha}{: x - number si m, alpha}
}
\value{
number -  Densitate de probabilitate in Rep. Pareto
}
\description{
Calculeaza densitatea de probabilitate in x a unei v.a. continue in repartitia Pareto - functie folosita in repartizarea grafica
}
