% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{e_densitate}
\alias{e_densitate}
\title{Densitate de probabilitate}
\usage{
e_densitate(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{f}{o repartitie de v.a. continua}
}
\value{
Bool
}
\description{
Pt o functie f, intodusa de utlizator se determina daca aceasta este densitate de probabilitate
}
