% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{genereaza}
\alias{genereaza}
\title{Genereaza n valori}
\usage{
genereaza(n,f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{n, f}{: f-o repartitie de v.a. continua, n cate valori sa genereze}
}
\value{
empty - afiseaza valorile generate
}
\description{
Calculeaza n valori dintr-o repartitie de v.a continua
}
