% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{info_exp}
\alias{info_exp}
\title{Fisa de sinteza (Rep. Exponentiala)}
\usage{
info_exp()
}
\value{
empty - afiseaza fisa de sinteza
}
\description{
Afisarea unei fise de sinteza pentru o v.a. continua in repartitia Exponentiala
}
