% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{info_norm}
\alias{info_norm}
\title{Fisa de sinteza (Rep. Normala)}
\usage{
info_norm()
}
\value{
empty - afiseaza fisa de sinteza
}
\description{
Afisarea unei fise de sinteza pentru o v.a. continua in repartitia Normala
}
