% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{info_Pareto}
\alias{info_Pareto}
\title{Fisa de sinteza (Rep. Pareto)}
\usage{
info_Pareto()
}
\value{
empty - afiseaza fisa de sinteza
}
\description{
Afisarea unei fise de sinteza pentru o v.a. continua in repartitia Pareto
}
