% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{info_uniform}
\alias{info_uniform}
\title{Fisa de sinteza (Rep. Uniforma)}
\usage{
info_unifrom()
}
\value{
empty - afiseaza fisa de sinteza
}
\description{
Afisarea unei fise de sinteza pentru o v.a. continua in repartitia Uniforma
}
