% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{marginale}
\alias{marginale}
\title{Densitatea marginala pentru 2 v.a. continue}
\usage{
marginale(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)}, a, b, c, d)
}
\arguments{
\item{f, a, b, c, d}{: f-o repartitie de v.a. continua, \link{a,b}, \link{c,d} - intevalele de definitie ale celor 2 v.a continue}
}
\value{
number - denistatea marginala
}
\description{
Realizeaza denistatea comuna a 2 v.a. continue, apoi calculeaza densitatea marginala pe baza ei
}
