% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{media_exp}
\alias{media_exp}
\title{Media (Rep. Exponentiala)}
\usage{
media_exp(x,lambda)
}
\arguments{
\item{x, lamda}{: x - number si lambda}
}
\value{
number - Media in Rep. Exponentiala
}
\description{
Calculeaza functia de medie in x a unei v.a. continue in repartitia Exponentiala - functie folosita in repartizarea grafica
}
