% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{media_norm}
\alias{media_norm}
\title{Media (Rep. Normala)}
\usage{
media_exp(x,miu, sigma)
}
\arguments{
\item{x, miu, sigma_p}{: x - number si miu, sigma}
}
\value{
number - Media in Rep. Normala
}
\description{
Calculeaza media unei v.a. continue in repartitia Normala - functie folosita in repartizarea grafica
}
