% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{media_pareto}
\alias{media_pareto}
\title{Media (Rep. Pareto)}
\usage{
media_exp(x,m,alpha)
}
\arguments{
\item{x, m, alpha}{: x - number si m,alpha}
}
\value{
number - Media in Rep. Pareto
}
\description{
Calculeaza media unei v.a. continue in repartitia Pareto - functie folosita in repartizarea grafica
}
