% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{media_unif}
\alias{media_unif}
\title{Media (Rep. Uniforma)}
\usage{
media_unif(x,a,b)
}
\arguments{
\item{x, a, b}{: x - number, \link{a,b} capetele intevalului de definitie pt X}
}
\value{
number - Media in Rep. Unfiorma
}
\description{
Calculeaza functia de medie in x a unei v.a. continue in repartitia Unfiorma - functie folosita in repartizarea grafica
}
