% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{mediana_exp}
\alias{mediana_exp}
\title{Mediana (Rep. Exponentiala)}
\usage{
mediana_exp(x,lambda)
}
\arguments{
\item{x, lamda}{: x - number si lambda}
}
\value{
number - Mediana in Rep. Exponentiala
}
\description{
Calculeaza mediana unei v.a. continue in repartitia Exponentiala - functie folosita in repartizarea grafica
}
