% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{mediana_norm}
\alias{mediana_norm}
\title{Mediana (Rep. Normala)}
\usage{
mediana_norm(x,lambda, miu, sigma)
}
\arguments{
\item{x, miu, sigma_p}{: x - number si miu, sigma}
}
\value{
number - Mediana in Rep. Normala
}
\description{
Calculeaza mediana unei v.a. continue in repartitia Normala - functie folosita in repartizarea grafica
}
