% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{mediana_pareto}
\alias{mediana_pareto}
\title{Mediana (Rep. Pareto)}
\usage{
mediana_pareto(x,m,alpha)
}
\arguments{
\item{x, m, alpha}{: x - number si m,alpha}
}
\value{
number - Mediana in Rep. Pareto
}
\description{
Calculeaza mediana unei v.a. continue in repartitia Pareto - functie folosita in repartizarea grafica
}
