% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{mediana_unif}
\alias{mediana_unif}
\title{Mediana (Rep. Uniforma)}
\usage{
mediana_unif(x,a,b)
}
\arguments{
\item{x, a, b}{: x - number, \link{a,b} capetele intevalului de definitie pt X}
}
\value{
number - Mediana in Rep. Uniforma
}
\description{
Calculeaza mediana unei v.a. continue in repartitia Uniforma - functie folosita in repartizarea grafica
}
