% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{medie}
\alias{medie}
\title{Media unei v.a continue X}
\usage{
medie(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{f}{: f-o repartitie de v.a. continua}
}
\value{
number - media
}
\description{
Calculeaza media unei v.a continue X
}
