% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{medie2}
\alias{medie2}
\title{Media unei v.a continue X trecuta print-o functie G}
\usage{
medie2(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)}, g(x){return (x+2)})
}
\arguments{
\item{f, g}{: f-o repartitie de v.a. continua, g functie continua}
}
\value{
number - media
}
\description{
Calculeaza media unei v.a continue X trecuta print-o functie G continua
}
