% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{moment_centrat}
\alias{moment_centrat}
\title{Momente centrat al unei v.a continue X}
\usage{
moment_centrat(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)}, k)
}
\arguments{
\item{f, k}{: f-o repartitie de v.a. continua, k - ordinul momentului centrat}
}
\value{
number - moment centrat
}
\description{
Calculeaza moment centrat de ordin k al unei v.a continue X
}
