% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{moment_centrat4}
\alias{moment_centrat4}
\title{Momente centrate ale unei v.a continue X}
\usage{
moment_centrat4(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{f:}{f-o repartitie de v.a. continua}
}
\value{
empty - afiseaza momentele centrale daca exista, daca nu afiseaza un mesaj corespunzator
}
\description{
Calculeaza momentele centrale pana la ordin 4 ale unei v.a continue X pe baza functiei moment_centrat
}
