% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{moment_initial}
\alias{moment_initial}
\title{Moment initial al unei v.a continue X}
\usage{
moment_initial (f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{f:}{f-o repartitie de v.a. continua}
}
\value{
number - moment_initial
}
\description{
Calculeaza momentul initial al unei v.a continue X
}
