% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{multiply}
\alias{multiply}
\title{Produsul a 2 functii}
\usage{
multiply(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)}, g(x){return (x+2)})
}
\arguments{
\item{f, g}{: f-o repartitie de v.a. continua, g functie continua}
}
\value{
function
}
\description{
Calculeaza produsul pentru 2 functii
}
