% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{plot2}
\alias{plot2}
\title{Grafic pentru repartitia uniforma a unei v.a. continue}
\usage{
plot2(f(x){if (x > a && x< b) return(1/(b-a)) else return(0)})
}
\arguments{
\item{f}{o repartitie de v.a. continua}
}
\value{
empty
}
\description{
Deseneaza graficul
}
