% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{r_exp}
\alias{r_exp}
\title{Repartitia (Rep. Exponentiala)}
\usage{
r_exp(x, lambda)
}
\arguments{
\item{x, lamda}{: x - number si lambda}
}
\value{
number - Repartitia in Rep. Exponentiala
}
\description{
Calculeaza functia de repatitie in x a unei v.a. continue in repartitia Exponentiala - functie folosita in repartizarea grafica
}
