% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{rpareto}
\alias{rpareto}
\title{Repartitia (Rep. Pareto)}
\usage{
rpareto(x,m,alpha)
}
\arguments{
\item{x, m, alpha}{: x - number si m,alpha}
}
\value{
number - Repartitia in Rep. Pareto
}
\description{
Calculeaza functia de repatitie in x a unei v.a. continue in repartitia Pareto - functie folosita in repartizarea grafica
}
