% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{var_exp}
\alias{var_exp}
\title{Variatia (Rep. Exponentiala)}
\usage{
val_exp(x,lambda)
}
\arguments{
\item{x, lamda}{: x - number si lambda}
}
\value{
number - Variatia in Rep. Exponentiala
}
\description{
Calculeaza variatia unei v.a. continue in repartitia Exponentiala - functie folosita in repartizarea grafica
}
