% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{var_norm}
\alias{var_norm}
\title{Variatia (Rep. Normala)}
\usage{
val_norm(x,miu,sigma_p)
}
\arguments{
\item{x, miu, sigma_p}{: x - number si miu, sigma}
}
\value{
number - Variatia in Rep. Normala
}
\description{
Calculeaza variatia unei v.a. continue in repartitia Normala - functie folosita in repartizarea grafica
}
