% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{var_pareto}
\alias{var_pareto}
\title{Variatia (Rep. Pareto)}
\usage{
val_pareto(x,m,alpha)
}
\arguments{
\item{x, m, alpha}{: x - number si m,alpha}
}
\value{
number - Variatia in Rep. Pareto
}
\description{
Calculeaza variatia unei v.a. continue in repartitia Pareto - functie folosita in repartizarea grafica
}
