% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{var_unif}
\alias{var_unif}
\title{Variatia (Rep. Uniforma)}
\usage{
val_unif(x,a,b)
}
\arguments{
\item{x, a, b}{: x - number, \link{a,b} capetele intevalului de definitie pt X}
}
\value{
number - Variatia in Rep. Uniforma
}
\description{
Calculeaza variatia unei v.a. continue in repartitia Unifroma - functie folosita in repartizarea grafica
}
